Create CombinedCode
Mode Enumeration: We use an enumeration to define the two modes (TEMPERATURE and MINIGAME).
Mode Buttons: Two buttons are added to switch between the modes.
displayTemperature Function: Displays the temperature on the LCD.
playMinigame Function: Allows the user to move a character on the LCD using the movement buttons.
Mode Checking in loop: The main loop checks which mode is active and calls the corresponding function.
Summary:
This setup allows you to switch between displaying the temperature and playing the minigame using the additional buttons. Make sure to verify the pin connections and ensure that the pull-up resistors are configured correctly.